requirements.txt
absl-py==2.4.0
agate==1.9.1
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asn1crypto==1.5.1
asttokens==3.0.1
astunparse==1.6.3
async-lru==2.1.0
attrs==25.4.0
babel==2.18.0
backports.tarfile==1.2.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
boto3==1.42.41
botocore==1.42.41
cachetools==6.2.6
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.4
cycler==0.12.1
daff==1.4.2
dbt-adapters==1.22.5
dbt-common==1.37.2
dbt-core==1.11.2
dbt-extractor==0.6.0
dbt-protos==1.0.427
dbt-semantic-interfaces==0.9.0
dbt-snowflake==1.11.1
debugpy==1.8.20
decorator==5.2.1
deepdiff==8.6.1
defusedxml==0.7.1
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.3
flatbuffers==25.12.19
fonttools==4.61.1
fqdn==1.5.1
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
greenlet==3.3.1
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
ipykernel==7.1.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isodate==0.7.2
isoduration==20.11.0
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
keras==3.13.2
keyring==25.7.0
kiwisolver==1.4.9
lark==1.3.1
leather==0.4.1
libclang==18.1.1
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mashumaro==3.14
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
ml_dtypes==0.5.4
more-itertools==10.8.0
msgpack==1.1.2
namex==0.1.0
narwhals==2.16.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
notebook==7.5.3
notebook_shim==0.2.4
numpy==2.4.2
opencv-python==4.13.0.90
opt_einsum==3.4.0
optree==0.18.0
orderly-set==5.5.0
overrides==7.7.0
packaging==26.0
pandas==2.3.3
pandocfilters==1.5.1
parsedatetime==2.6
parso==0.8.5
pathspec==0.12.1
pillow==12.1.0
platformdirs==4.5.1
prometheus_client==0.24.1
prompt_toolkit==3.0.52
protobuf==6.33.5
psutil==7.2.2
pure_eval==0.2.3
pyarrow==23.0.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.11.0
pyOpenSSL==25.3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
pywin32-ctypes==0.2.3
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.2
rpds-py==0.30.0
s3transfer==0.16.0
scikit-learn==1.8.0
scipy==1.17.0
Send2Trash==2.1.0
six==1.17.0
smmap==5.0.2
snowflake-connector-python==4.2.0
snowflake-sqlalchemy==1.8.2
snowplow-tracker==1.1.0
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==2.0.46
sqlparse==0.5.4
stack-data==0.6.3
streamlit==1.53.1
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
tomlkit==0.14.0
tornado==6.5.4
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
watchdog==6.0.0
wcwidth==0.5.3
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.5
widgetsnbextension==4.0.15
wrapt==2.1.0
zipp==3.23.0
